requirements.txt
numpy
libigl
scipy